borrado de comentarios

+++ videoclub/app/Http/Controllers/UserImageController.php
@@ -8,32 +8,36 @@ use App\Models\Admin;
 use App\Models\Free;
 use App\Models\Premium;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Http\UploadedFile;
+use Illuminate\Support\Facades\Validator;
 
 class UserImageController extends Controller
 {
 
     public function update(Request $request, $id)
     {
-        // dd($request->all());
         $user = User::findOrFail($id);
 
         $attributes = $request->validate([
-        'image' => 'file|nullable'
+           'image' => 'image|nullable'
         ]);
 
-        if($request->hasFile('image')){
+        if(!$request->hasFile('image')){
+            return response()->json('error'); //422
+        }
 
+        if($request->hasFile('image')){
             if(isset($user->image)){
                 $old_image = $user->image;
                 Storage::disk('public')->delete($old_image);
             }
 
             $filename = $request->file('image')->getClientOriginalName();
-            $attributes['image'] = request()->file('image')->storeAs('user_profile_img', $filename, 'public',);
+            $attributes['image'] = request()->file('image')->storeAs('user_profile_img', $filename, 'public');
         }
 
         $user->update($attributes);
-        // dd($user);
-        return response()->json($user);;
+
+        return response()->json($user);
     }
 }
